Update bitshuffle to latest version (with lz4 support)
